rename to markers view
